refactor: integration test modules
