+++ rust/signed_doc/tests/common/mod.rs
@@ -1,3 +1,5 @@
+#![allow(dead_code)]
+
 use std::{collections::HashMap, str::FromStr};
 
 use catalyst_signed_doc::*;
